feat(cli): improve output and continue upgrading packages on errors
Improve the output of `go-global-update`. Include icons in the output
and skip some output messages to make it less noisy. Include current
version of modules.

Refactor the logic to continue upgrading packages if one of them fails.
The exit code will be 1 at the end anyway, signaling that the some
upgrade failed. The error messages are retained for failed upgrades.