feat(cli): skip updating binaries built from source
Detect when a binary was built or installed from source (either using
`go build main.go` on a Go file specifically, or using `go install` in
the cloned repository) and skip updating such binaries.

Binaries installed from source (`go install` in the cloned repository)
may have been modified by the engineers prior to being built, which
means their version may not be the same as the latest versions in the
upstream repositories.

On top of that, binaries built from a Go file directly (`go build
main.go`) have `command-line-arguments` set as their `path` in `go
version -m`. This makes it impossible to see where a binary came from.`

In both cases, a message is logged to the console and the command still
passes. Encountering such binaries does not cause `go-global-update` to
fail.

Closes #3