fix(cli): support go 1.18 command outputs
The output of `go version -m binary-name` is different on go 1.18 for
binaries built using `go build`. The `mod` line is missing (not printed
at all).

This commit changes the parsing logic to handle this situation and avoid
attempting to update or check the latest version of binaries built from
source.

This should add support for using go-global-update on go 1.18.