feat: add a --force flag to force-reinstall all packages
This commit introduces a new `--force` flag (alias: -f). It overrides
the upgrade behavior to force reinstall binaries when they do not have a
new version available.

The option can be used when a new version of go is released that fixes
vulnerabilities and the user wants to reinstall all the binaries using
the new version.

Closes https://github.com/Gelio/go-global-update/issues/22